+++ next.config.ts
@@ -4,7 +4,6 @@ const nextConfig: NextConfig = {
   allowedDevOrigins: [
     "127.0.0.1",
   ],
-  /* config options here */
   reactCompiler: true,
 };